feat(mantaray): flip the crate no_std, seam rand, and fence the legacy surface
The crate carries the no_std attribute with the trie modules riding the
std feature until the primitives core builds bare-metal; the bare no_std
surface (metadata keys and the obfuscation key) is gated on the rv64 and
wasm32 lanes. Random obfuscation-key generation moves behind a dedicated
rand feature so std builds without an entropy source. The lazy-trie
Manifest, ManifestIter, and ManifestBuilder move to legacy/ with
deprecation notes and a migration table; the traversal and save
machinery serving only that surface follows them, while the add, remove,
and load substrate stays with the editor. Removal of the legacy surface
is gated on the manifest 1.0 key-value store replacing it.
